+++ cli/tools/test/reporters/pretty.rs
@@ -1,5 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
+use super::fmt::format_test_error;
+use super::fmt::to_relative_path_or_remote_url;
 use super::*;
 
 pub struct PrettyTestReporter {
@@ -39,7 +41,7 @@ impl PrettyTestReporter {
         "{}",
         colors::gray(format!(
           "{} => ",
-          self.to_relative_path_or_remote_url(&description.origin)
+          to_relative_path_or_remote_url(&self.cwd, &description.origin)
         ))
       );
     }
@@ -48,19 +50,6 @@ impl PrettyTestReporter {
     // flush for faster feedback when line buffered
     std::io::stdout().flush().unwrap();
     self.scope_test_id = Some(description.id);
-  }
-
-  fn to_relative_path_or_remote_url(&self, path_or_url: &str) -> String {
-    let url = Url::parse(path_or_url).unwrap();
-    if url.scheme() == "file" {
-      if let Some(mut r) = self.cwd.make_relative(&url) {
-        if !r.starts_with("../") {
-          r = format!("./{r}");
-        }
-        return r;
-      }
-    }
-    path_or_url.to_string()
   }
 
   fn force_report_step_wait(&mut self, description: &TestStepDescription) {
@@ -174,7 +163,7 @@ impl PrettyTestReporter {
       &desc.name,
       colors::gray(format!(
         "=> {}:{}:{}",
-        self.to_relative_path_or_remote_url(&desc.location.file_name),
+        to_relative_path_or_remote_url(&self.cwd, &desc.location.file_name),
         desc.location.line_number,
         desc.location.column_number
       ))
@@ -193,7 +182,7 @@ impl PrettyTestReporter {
       long_name,
       colors::gray(format!(
         "=> {}:{}:{}",
-        self.to_relative_path_or_remote_url(&desc.location.file_name),
+        to_relative_path_or_remote_url(&self.cwd, &desc.location.file_name),
         desc.location.line_number,
         desc.location.column_number
       ))
@@ -216,7 +205,7 @@ impl TestReporter for PrettyTestReporter {
         "running {} {} from {}",
         plan.total,
         inflection,
-        self.to_relative_path_or_remote_url(&plan.origin)
+        to_relative_path_or_remote_url(&self.cwd, &plan.origin)
       ))
     );
     self.in_new_line = true;
@@ -314,7 +303,7 @@ impl TestReporter for PrettyTestReporter {
     }
     println!(
       "Uncaught error from {} {}",
-      self.to_relative_path_or_remote_url(origin),
+      to_relative_path_or_remote_url(&self.cwd, origin),
       colors::red("FAILED")
     );
     self.in_new_line = true;
@@ -366,7 +355,7 @@ impl TestReporter for PrettyTestReporter {
         "{} {} ...",
         colors::gray(format!(
           "{} =>",
-          self.to_relative_path_or_remote_url(&desc.origin)
+          to_relative_path_or_remote_url(&self.cwd, &desc.origin)
         )),
         self.format_test_step_ancestry(desc, tests, test_steps)
       );
@@ -434,7 +423,7 @@ impl TestReporter for PrettyTestReporter {
         if let Some(js_error) = uncaught_error {
           let failure_title = format!(
             "{} (uncaught error)",
-            self.to_relative_path_or_remote_url(&origin)
+            to_relative_path_or_remote_url(&self.cwd, &origin)
           );
           println!("{}", &failure_title);
           println!(
